+++ src/model/ContainId.java
@@ -8,18 +8,18 @@ import javax.persistence.Embeddable;
 public class ContainId implements Serializable {
 
     private int idOrder;
-    private int isbnBook;
+    private long isbnBook;
 
     public ContainId() {
     }
 
-    public ContainId(int idOrder, int isbnBook) {
+    public ContainId(int idOrder, long isbnBook) {
         this.idOrder = idOrder;
         this.isbnBook = isbnBook;
     }
 
     // Getters, Setters, Equals y HashCode (OBLIGATORIOS)
-    public int getIdOrder() {
+    public long getIdOrder() {
         return idOrder;
     }
 
@@ -27,11 +27,11 @@ public class ContainId implements Serializable {
         this.idOrder = idOrder;
     }
 
-    public int getIsbnBook() {
+    public long getIsbnBook() {
         return isbnBook;
     }
 
-    public void setIsbnBook(int isbnBook) {
+    public void setIsbnBook(long isbnBook) {
         this.isbnBook = isbnBook;
     }
